requirements.txt
# ====================================
# Voice Cloning Platform - Requirements
# Python 3.10+
# Model removed - Using External API
# ====================================

# Core Django Framework
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
django-cors-headers==4.3.1
django-storages==1.14.2
gunicorn==21.2.0
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment
python-decouple==3.8

# Payment Gateways
stripe==7.4.0
paypalrestsdk==1.13.3

# Authentication & Security
PyJWT==2.10.1
oauthlib==3.3.1
requests-oauthlib==2.0.0
python3-openid==3.2.0
cryptography==46.0.3
pyOpenSSL==25.3.0

# Audio Processing (Basic - for file handling only)
soundfile==0.12.1
pydub==0.25.1

# Machine Learning Utils (Basic)
numpy==1.26.4

# API & Web Framework
uvicorn==0.38.0
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
python-multipart==0.0.20

# Async & Networking
aiohttp==3.13.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
async-timeout==5.0.1
websockets==15.0.1
multidict==6.7.0
yarl==1.22.0
frozenlist==1.8.0
propcache==0.4.1

# Task Queue & Caching
celery==5.3.4
redis==5.0.1
amqp==5.3.1
billiard==4.2.2
kombu==5.6.0
vine==5.1.0

# Cloud Storage (AWS S3)
boto3==1.34.0
botocore==1.34.162
s3transfer==0.9.0
jmespath==1.0.1

# Google Cloud Storage
google-cloud-storage==2.19.0
google-cloud-core==2.5.0
google-auth==2.43.0
google-api-core==2.28.1
google-resumable-media==2.7.2
google-crc32c==1.7.1
googleapis-common-protos==1.72.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1

# Monitoring & Logging
sentry-sdk==2.43.0
loguru==0.7.3

# Utilities
requests==2.32.5
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
tqdm==4.67.1
rich==13.9.4
Pygments==2.19.2
markdown-it-py==4.0.0
mdurl==0.1.2
wcwidth==0.2.14

# Configuration & Serialization
PyYAML==6.0.3
tomlkit==0.13.3
tomli==2.3.0
pydantic==2.10.6
pydantic_core==2.27.2
annotated-types==0.7.0

# Data Processing
orjson==3.11.4

# Template & Text Processing
Jinja2==3.1.6
MarkupSafe==2.1.5
regex==2025.11.3

# Visualization (for admin charts)
matplotlib==3.10.7
contourpy==1.3.2
cycler==0.12.1
fonttools==4.60.1
kiwisolver==1.4.9
Pillow==10.1.0

# File & Path Utilities
filelock==3.13.4
fsspec==2025.9.0
platformdirs==4.5.0
cachetools==6.2.1
shellingham==1.5.4

# System & Process
psutil==7.1.3

# CLI & Terminal
typer==0.20.0
typer-slim==0.20.0
prompt_toolkit==3.0.52

# Misc
anyio==4.11.0
sniffio==1.3.1
exceptiongroup==1.3.0
attrs==25.4.0
decorator==5.2.1
six==1.17.0
packaging==25.0
pyparsing==3.2.5

# Django Specific
asgiref==3.10.0
sqlparse==0.5.3
pytz==2025.2
tzdata==2025.2
python-dateutil==2.9.0.post0
defusedxml==0.7.1

# Development & Testing
ruff==0.14.4

# Additional Dependencies
brotli==1.2.0
cffi==2.0.0
pycparser==2.23
typing_extensions==4.15.0
